test: 기존의 ExpressionParser test를 Formula test로 변경
- Formula 객체를 생성하여 test하는 코드로 변경
